ytdlp_service/requirements.txt
fastapi==0.112.1
pydantic==2.8.2
python-dotenv==1.0.1
Requests==2.32.3
Telethon==1.36.0
uvicorn==0.30.6
